go: create a separate package for client
