+++ go/ndt7-client/client/client_test.go
@@ -1,4 +1,4 @@
-package main
+package client
 
 import (
 	"errors"
